requirements.txt
aiogram~=3.23.0
yt-dlp==2025.1.26
pytest>=8.0
pytest-asyncio>=0.23
python-dotenv~=1.2.1
pip~=25.3
typing_extensions~=4.15.0
attrs~=25.4.0
rsa~=4.9.1
pyasn1~=0.6.1
idna~=3.11
multidict~=6.7.0
propcache~=0.4.1
certifi~=2025.11.12
pluggy~=1.6.0
iniconfig~=2.3.0
pydantic~=2.9.2
aiofiles~=24.1.0
aiohttp~=3.10.11
yarl~=1.22.0
aiosignal~=1.4.0
frozenlist~=1.8.0
aiohappyeyeballs~=2.6.1
pydantic_core~=2.23.4
pyaes~=1.6.1
packaging~=25.0